feat: 날짜 포맷 함수 위치 수정 (#37)
